Add shift utility methods toUnisgnedInt, toInt, parseSensorAdvertisement to utilities file
